Add files via upload
This repository contains the Credit Card Dataset for Machine Learning, which is used for analyzing and predicting credit card transactions. The dataset includes features related to customer transactions, with a focus on fraud detection. Pre-processing steps such as handling missing data, encoding categorical variables, and scaling numerical features have been applied. The project uses Python, with libraries like pandas and numpy for data manipulation, and is designed for use in machine learning models.